updated the sharing registry apis

+++ custos-sharing-registry-service/sharing-service-core/src/main/java/org/apache/custos/sharing/service/core/models/GroupMembership.java
@@ -6,8 +6,8 @@ public class GroupMembership {
     private String childId;
     private String domainId;
     private GroupChildType childType;
-    private double createdTime;
-    private double updatedTime;
+    private Long createdTime;
+    private Long updatedTime;
 
     public String getParentId() {
         return parentId;
@@ -41,19 +41,19 @@ public class GroupMembership {
         this.childType = childType;
     }
 
-    public double getCreatedTime() {
+    public Long getCreatedTime() {
         return createdTime;
     }
 
-    public void setCreatedTime(double createdTime) {
+    public void setCreatedTime(Long createdTime) {
         this.createdTime = createdTime;
     }
 
-    public double getUpdatedTime() {
+    public Long getUpdatedTime() {
         return updatedTime;
     }
 
-    public void setUpdatedTime(double updatedTime) {
+    public void setUpdatedTime(Long updatedTime) {
         this.updatedTime = updatedTime;
     }
